\name{eigenCentrality}
\alias{eigenCentrality}
\title{
  Calculates the eigen centrality of a network.
}
\description{
  Calculates the eigen centrality of a network.
}
\usage{
eigenCentrality(x='matrix')
}
\arguments{
  \item{x}{
    A matrix defining a network graph.
  }
}

\value{
  Returns the eigen based centrality of the network.
}

\references{
  Bonacich, P., 1987. Power and centrality: a family of measures. American
  Journal of Sociology 92: 1170-1182.
}

\author{
Stuart R. Borrett (borretts@uncw.edu)
Matthew K. Lau (mkl48@nau.edu)
}

